LayoutTests/http/tests/security/window-named-proto-expected.txt

CONSOLE MESSAGE: line 2: TypeError: null is not an object (evaluating 'document.body.innerHTML')
